wip k3d, silly go mod

+++ test/e2e/common.go
@@ -12,6 +12,7 @@ import (
 	"testing"
 	"time"
 
+	// k3dcluster "github.com/rancher/k3d/v5/cmd/cluster"
 	"github.com/stretchr/testify/require"
 	v1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
@@ -39,7 +40,7 @@ type ZarfE2ETest struct {
 }
 
 func getKubeconfigPath() (string, error) {
-
+	// k3dcluster.NewCmdClusterCreate()
 	// Check if the $KUBECONFIG env is set
 	// TODO: It would probably be good to verify a useable kubeconfig lives here
 	kubeconfigEnv := os.Getenv("KUBECONFIG")
@@ -62,7 +63,7 @@ func getKubeconfigPath() (string, error) {
 	return kubeconfigPath, err
 }
 
-func (e2e *ZarfE2ETest) setUp() error {
+func (e2e *ZarfE2ETest) setUpKind() error {
 
 	// Determine what the name of the zarfBinary should be
 	e2e.zarfBinPath = path.Join("../../build", getCLIName())
@@ -102,7 +103,6 @@ func (e2e *ZarfE2ETest) setUp() error {
 	}
 
 	// Wait for the cluster to have pods before we let the test suite run
-	// TODO: Pretty sure there's a cleaner way to do this
 	attempt := 0
 	for attempt < 10 {
 		pods, err := e2e.clientset.CoreV1().Pods("kube-system").List(context.TODO(), metav1.ListOptions{})
@@ -121,14 +121,72 @@ func (e2e *ZarfE2ETest) setUp() error {
 	return err
 }
 
-func (e2e *ZarfE2ETest) tearDown() error {
+func (e2e *ZarfE2ETest) tearDownKind() error {
+	if os.Getenv("SKIP_TEARDOWN") != "" || e2e.clusterAlreadyExists {
+		return nil
+	}
+
+	// k3dcluster.NewCmdClusterCreate()
+
 	// Delete the cluster and kubeconfig file
 	provider := cluster.NewProvider(cluster.ProviderWithLogger(kindcmd.NewLogger()))
 	err := provider.Delete(e2e.clusterName, e2e.kubeconfigPath)
 	os.Remove(e2e.kubeconfigPath)
-
 	return err
 }
+
+// func (e2e *ZarfE2ETest) setUpK3D() error {
+
+// 	// Determine what the name of the zarfBinary should be
+// 	e2e.zarfBinPath = path.Join("../../build", getCLIName())
+
+// 	var err error
+// 	// Create or get the kubeconfig
+// 	e2e.kubeconfigPath, err = getKubeconfigPath()
+// 	if err != nil {
+// 		return err
+// 	}
+
+// 	createClusterCommand := k3dcluster.NewCmdClusterCreate()
+// 	err = createClusterCommand.ExecuteContext(context.TODO())
+// 	if err != nil {
+// 		fmt.Println("ERROR WHEN TRYING TO SET UP K3D CLUSTER")
+// 		return err
+// 	}
+
+// 	// Get config and client for the k8s cluster
+// 	e2e.restConfig, err = clientcmd.BuildConfigFromFlags("", e2e.kubeconfigPath)
+// 	if err != nil {
+// 		return err
+// 	}
+// 	e2e.clientset, err = kubernetes.NewForConfig(e2e.restConfig)
+// 	if err != nil {
+// 		return err
+// 	}
+
+// 	// Wait for the cluster to have pods before we let the test suite run
+// 	attempt := 0
+// 	for attempt < 10 {
+// 		pods, err := e2e.clientset.CoreV1().Pods("kube-system").List(context.TODO(), metav1.ListOptions{})
+// 		if err == nil && len(pods.Items) >= 0 {
+// 			fmt.Printf("💥 Cluster %s ready. You can access it by setting:\nexport KUBECONFIG='%s'\n", e2e.clusterName, e2e.kubeconfigPath)
+// 			break
+// 		}
+
+// 		time.Sleep(1 * time.Second)
+// 		attempt++
+// 		if attempt > 15 {
+// 			return errors.New("unable to connect to KinD cluster for e2e tests")
+// 		}
+// 	}
+// 	return nil
+// }
+
+// func (e2e *ZarfE2ETest) tearDownK3D() error {
+// 	deleteClusterCommand := k3dcluster.NewCmdClusterDelete()
+// 	err := deleteClusterCommand.ExecuteContext(context.TODO())
+// 	return err
+// }
 
 func getCLIName() string {
 	var binaryName string
@@ -145,11 +203,9 @@ func getCLIName() string {
 }
 
 func (e2e *ZarfE2ETest) cleanupAfterTest(t *testing.T) {
-	fmt.Println("Test is finished, cleaning up now")
-
 	// Use Zarf to perform chart uninstallation
-	_, err := exec.Command(e2e.zarfBinPath, "destroy", "--confirm", "--remove-components", "-l=trace").CombinedOutput()
-	require.NoError(t, err, "unable to destroy the zarf cluster when cleaning up after a test")
+	output, err := e2e.execZarfCommand("destroy", "--confirm", "--remove-components", "-l=trace")
+	require.NoError(t, err, output)
 
 	// Remove files created for the test
 	for _, filePath := range e2e.filesToRemove {
@@ -166,11 +222,6 @@ func (e2e *ZarfE2ETest) cleanupAfterTest(t *testing.T) {
 		}
 	}
 	e2e.cmdsToKill = []*exec.Cmd{}
-
-	fmt.Println("sleeping for 10 seconds after clean up.. for reasons..")
-	time.Sleep(10 * time.Second)
-	fmt.Println("done sleeping!")
-
 }
 
 func (e2e *ZarfE2ETest) execCommandInPod(podname, namespace string, cmd []string) (string, string, error) {
@@ -189,7 +240,6 @@ func (e2e *ZarfE2ETest) execCommandInPod(podname, namespace string, cmd []string
 
 	exec, err := remotecommand.NewSPDYExecutor(e2e.restConfig, "POST", req.URL())
 	if err != nil {
-		fmt.Println("@JPERRY something was broken with the spdy executor...")
 		return "", "", err
 	}
 
@@ -202,9 +252,18 @@ func (e2e *ZarfE2ETest) execCommandInPod(podname, namespace string, cmd []string
 	return stdoutBuffer.String(), stderrBuffer.String(), err
 }
 
-func (e2e *ZarfE2ETest) execZarfCommand(commandString ...string) error {
-	cmd := exec.Command(e2e.zarfBinPath, commandString...)
-	// cmd.Stdout = os.Stdout
-	// cmd.Stderr = os.Stderr
-	return cmd.Run()
+// TODO: It might be a nice feature to read some flag/env and change the stdout and stderr to pipe to the terminal running the test
+func (e2e *ZarfE2ETest) execZarfCommand(commandString ...string) (string, error) {
+	output, err := exec.Command(e2e.zarfBinPath, commandString...).CombinedOutput()
+	return string(output), err
+}
+
+func (e2e *ZarfE2ETest) execZarfBackgroundCommand(commandString ...string) error {
+	// Create a tunnel to the git resources
+	tunnelCmd := exec.Command(e2e.zarfBinPath, commandString...)
+	err := tunnelCmd.Start()
+	e2e.cmdsToKill = append(e2e.cmdsToKill, tunnelCmd)
+	time.Sleep(1 * time.Second)
+
+	return err
 }
